+++ R/stat-stratum.r
@@ -100,7 +100,7 @@ StatStratum <- ggproto(
       axis_ind <- get_axes(names(data))
       data <- to_lodes(data = data, axes = axis_ind)
       # positioning requires numeric 'x'
-      data <- dplyr::arrange(data, x, stratum, alluvium)
+      data <- data[with(data, order(x, stratum, alluvium)), , drop = FALSE]
       data$x <- contiguate(data$x)
     }
     
@@ -135,10 +135,10 @@ StatStratum <- ggproto(
     # sort in preparation for calculating cumulative weights
     data <- if (is.na(decreasing)) {
       arr_fun <- if (reverse) dplyr::desc else identity
-      dplyr::arrange(data, PANEL, x, arr_fun(stratum))
+      data[with(data, order(PANEL, x, arr_fun(stratum))), , drop = FALSE]
     } else {
       arr_fun <- if (decreasing) dplyr::desc else identity
-      dplyr::arrange(data, PANEL, x, arr_fun(weight))
+      data[with(data, order(PANEL, x, arr_fun(weight))), , drop = FALSE]
     }
     
     # calculate cumulative weights