+++ inst/examples/ex-alluvial-data.r
@@ -11,10 +11,10 @@ head(titanic_lodes)
 is_alluvial(titanic_lodes,
             key = "x", value = "stratum", id = "alluvium",
             weight = "Freq")
-# again in lodes format, this time keeping 'Class' as a variable
+# again in lodes format, this time diffusing the 'Class' variable
 titanic_lodes2 <- to_lodes(titanic_alluvia,
                            key = "variable", value = "value", id = "passenger",
-                           axes = 1:3, keep = "Class")
+                           axes = 1:3, diffuse = "Class")
 head(titanic_lodes2)
 is_alluvial(titanic_lodes2,
             key = "variable", value = "value", id = "passenger",
@@ -27,9 +27,21 @@ is_alluvial(majors,
             key = "semester", value = "curriculum", id = "student",
             logical = FALSE)
 # curriculum data in alluvia format
-majors2 <- to_alluvia(majors,
-                      key = "semester", value = "curriculum", id = "student")
-head(majors2)
-is_alluvial(majors2,
+majors_alluvia <- to_alluvia(
+  majors,
+  key = "semester", value = "curriculum", id = "student"
+)
+head(majors_alluvia)
+is_alluvial(majors_alluvia,
             axes = 2:9,
             logical = FALSE)
+
+# distill variables that vary within 'id' values
+set.seed(1)
+majors$hypo_grade <- LETTERS[sample(5, size = nrow(majors), replace = TRUE)]
+majors_alluvia2 <- to_alluvia(
+  majors,
+  key = "semester", value = "curriculum", id = "student",
+  distill = "most"
+)
+head(majors_alluvia2)