+++ tests/testthat/test-is-alluvial.r
@@ -3,8 +3,8 @@ context("is_alluvial")
 
 titanic_alluvia <- as.data.frame(Titanic)
 
-test_that("is_alluvial recognizes alluvium-form Titanic data", {
-  expect_warning(is_alluvial(titanic_alluvia),
+test_that("is_alluvial recognizes alluvia-format Titanic data", {
+  expect_message(is_alluvial(titanic_alluvia),
                  "[Mm]issing")
   expect_true(is_alluvial(titanic_alluvia, axes = 1:4))
   expect_true(is_alluvial(titanic_alluvia, axes = c("Class", "Sex")))
@@ -21,14 +21,14 @@ titanic_lodes$Value <- factor(titanic_lodes$Value,
                               levels = do.call(c, lapply(titanic_alluvia[, 1:4],
                                                          levels)))
 
-test_that("is_alluvial recognizes lode-form Titanic data", {
-  expect_warning(is_alluvial(titanic_lodes),
+test_that("is_alluvial recognizes lodes-format Titanic data", {
+  expect_message(is_alluvial(titanic_lodes),
                  "[Mm]issing")
   expect_true(is_alluvial(titanic_lodes,
                           key = "Variable", value = "Value", id = "Index"))
   expect_true(is_alluvial(titanic_lodes,
                           key = "Variable", value = "Value", id = "Index",
                           weight = "Freq"))
-  expect_warning(is_alluvial(titanic_lodes, axes = 3:4),
+  expect_message(is_alluvial(titanic_lodes, axes = 3:4),
                  "[Mm]issing")
 })